+++ src/types/spawn-please.d.ts
@@ -1,1 +1,2 @@
+// add to tsconfig compilerOptions.paths
 declare module 'spawn-please'
